// src/web/shared/proto/jaiabot/messages/engineering.ts
// Code generated by protoc-gen-ts_proto. DO NOT EDIT.
// versions:
//   protoc-gen-ts_proto  v2.12.0
//   protoc               v3.21.12
// source: jaiabot/messages/engineering.proto

/* eslint-disable */
import type { Bounds } from "./bounds";
import type { Link } from "./link";
import type { BottomDepthSafetyParams } from "./mission";
import type { PamState } from "./pam";

export const protobufPackage = "jaiabot.protobuf";

export enum BotStatusRate {
    BotStatusRate_2_Hz = "BotStatusRate_2_Hz",
    BotStatusRate_1_Hz = "BotStatusRate_1_Hz",
    BotStatusRate_2_SECONDS = "BotStatusRate_2_SECONDS",
    BotStatusRate_5_SECONDS = "BotStatusRate_5_SECONDS",
    BotStatusRate_10_SECONDS = "BotStatusRate_10_SECONDS",
    BotStatusRate_20_SECONDS = "BotStatusRate_20_SECONDS",
    BotStatusRate_40_SECONDS = "BotStatusRate_40_SECONDS",
    BotStatusRate_60_SECONDS = "BotStatusRate_60_SECONDS",
    BotStatusRate_NO_RF = "BotStatusRate_NO_RF",
}

export interface PIDControl {
    timeout?: number | undefined;
    throttle?: number | undefined;
    speed?: PIDControl_PIDSettings | undefined;
    rudder?: number | undefined;
    heading?: PIDControl_PIDSettings | undefined;
    port_elevator?: number | undefined;
    stbd_elevator?: number | undefined;
    roll?: PIDControl_PIDSettings | undefined;
    pitch?: PIDControl_PIDSettings | undefined;
    depth?: PIDControl_PIDSettings | undefined;
    led_switch_on?: boolean | undefined;
    heading_constant?: PIDControl_PIDSettings | undefined;
}

export interface PIDControl_PIDSettings {
    target?: number | undefined;
    Kp?: number | undefined;
    Ki?: number | undefined;
    Kd?: number | undefined;
}

export interface GPSRequirements {
    transit_hdop_req?: number | undefined;
    transit_pdop_req?: number | undefined;
    after_dive_hdop_req?: number | undefined;
    after_dive_pdop_req?: number | undefined;
    transit_gps_fix_checks?: number | undefined;
    transit_gps_degraded_fix_checks?: number | undefined;
    after_dive_gps_fix_checks?: number | undefined;
}

export interface RFDisableOptions {
    rf_disable?: boolean | undefined;
    rf_disable_timeout_mins?: number | undefined;
}

export interface IMUCalibration {
    run_cal?: boolean | undefined;
}

export interface Pam {
    start_pam?: boolean | undefined;
    stop_pam?: boolean | undefined;
    pam_state?: PamState | undefined;
}

export interface Engineering {
    bot_id?: number | undefined;
    time?: string | undefined;
    pid_control?: PIDControl | undefined;
    query_engineering_status?: boolean | undefined;
    query_bot_status?: boolean | undefined;
    engineering_messages_enabled?: boolean | undefined;
    bot_status_rate?: BotStatusRate | undefined;
    gps_requirements?: GPSRequirements | undefined;
    rf_disable_options?: RFDisableOptions | undefined;
    bottom_depth_safety_params?: BottomDepthSafetyParams | undefined;
    imu_cal?: IMUCalibration | undefined;
    pam?: Pam | undefined;
    /** For User flagging of events */
    flag?: number | undefined;
    /** For calibration/configuration of Arduino motor/rudder bounds */
    bounds?: Bounds | undefined;
    /** Special case: not sent in message - written upon receipt based on the Link traversed */
    link?: Link | undefined;
}