// src/web/shared/proto/jaiabot/messages/jaia_dccl.ts
export enum Command_CommandType {
    /** MISSION_PLAN - pre mission */
    MISSION_PLAN = "MISSION_PLAN",
    ACTIVATE = "ACTIVATE",
    START_MISSION = "START_MISSION",
    MISSION_PLAN_FRAGMENT = "MISSION_PLAN_FRAGMENT",
    /** NEXT_TASK - during any mission */
    NEXT_TASK = "NEXT_TASK",
    RETURN_TO_HOME = "RETURN_TO_HOME",
    STOP = "STOP",
    PAUSE = "PAUSE",
    RESUME = "RESUME",
    /** REMOTE_CONTROL_SETPOINT - during remote control mission */
    REMOTE_CONTROL_SETPOINT = "REMOTE_CONTROL_SETPOINT",
    REMOTE_CONTROL_TASK = "REMOTE_CONTROL_TASK",
    REMOTE_CONTROL_RESUME_MOVEMENT = "REMOTE_CONTROL_RESUME_MOVEMENT",
    /** RECOVERED - post mission */
    RECOVERED = "RECOVERED",
    SHUTDOWN = "SHUTDOWN",
    RETRY_DATA_OFFLOAD = "RETRY_DATA_OFFLOAD",
    DATA_OFFLOAD_COMPLETE = "DATA_OFFLOAD_COMPLETE",
    DATA_OFFLOAD_FAILED = "DATA_OFFLOAD_FAILED",
    /** RESTART_ALL_SERVICES - debugging low level commands */
    RESTART_ALL_SERVICES = "RESTART_ALL_SERVICES",
    REBOOT_COMPUTER = "REBOOT_COMPUTER",
    SHUTDOWN_COMPUTER = "SHUTDOWN_COMPUTER",
}