// src/web/shared/proto/jaiabot/messages/mission.ts
export enum MissionTask_TaskType {
    NONE = "NONE",
    DIVE = "DIVE",
    STATION_KEEP = "STATION_KEEP",
    SURFACE_DRIFT = "SURFACE_DRIFT",
    CONSTANT_HEADING = "CONSTANT_HEADING",
}

export interface MissionTask_DiveParameters {
    max_depth?: number | undefined;
    depth_interval?: number | undefined;
    hold_time?: number | undefined;
    bottom_dive?: boolean | undefined;
}

export interface MissionTask_DriftParameters {
    drift_time?: number | undefined;
}